Conflicts:
	.gitignore
	grails-app/conf/UnifinaCorePluginResources.groovy

+++ grails-app/controllers/com/unifina/controller/data/StreamController.groovy
@@ -159,16 +159,18 @@ class StreamController {
 	}
 	
 	def search() {
-		Set<Feed> allowedFeeds = springSecurityService.currentUser?.feeds ?: new HashSet<>()
 		List<Map> streams = []
+		SecUser user = springSecurityService.currentUser
+		Set<Feed> allowedFeeds = user.feeds ?: new HashSet<>()
 
 		if (!allowedFeeds.isEmpty()) {
 			String hql = "select new map(s.id as id, s.name as name, s.feed.module.id as module, s.description as description) from Stream s "+
 				"left outer join s.feed "+
 				"left outer join s.feed.module "+
-				"where (s.name like '"+params.term+"%' or s.description like '%"+params.term+"%') "
-				"and s.feed.id in ("+allowedFeeds.collect{ feed -> feed.id }.join(',')+") "
-
+				"where (s.name like '"+params.term+"%' or s.description like '%"+params.term+"%') "+
+				"and s.feed.id in ("+allowedFeeds.collect{ feed -> feed.id }.join(',')+") "+
+				"and (s.feed.id != 7 OR s.user.id = ${user.id}) " // Quick fix for CORE-452, needs proper ACL
+				
 				if (params.feed) {
 					hql += " and s.feed.id="+Feed.load(params.feed).id
 				}
@@ -203,8 +205,11 @@ class StreamController {
 			MultipartFile file = request.getFile("file")
 			temp = File.createTempFile("csv_upload_", ".csv")
 			file.transferTo(temp)
-			
-			CSVImporter csv = new CSVImporter(temp)
+
+			Map config = (stream.streamConfig ? JSON.parse(stream.streamConfig) : [:])
+			List fields = config.fields ? config.fields : []
+
+			CSVImporter csv = new CSVImporter(temp, fields)
 			if (csv.getSchema().timestampColumnIndex==null) {
 				flash.message = "Unfortunately we couldn't recognize some of the fields in the CSV-file. But no worries! With a couple of confirmations we still can import your data."
 				response.status = 500
@@ -216,11 +221,11 @@ class StreamController {
 				render ([success:true] as JSON)
 			}
 		} catch (Exception e) {
-				flash.message = "An error occurred while handling file: $e"
+				log.error("Failed to import file", e)
 				response.status = 500
-				render ([success:false, error: e.toString()] as JSON)
+				render ([success:false, error: e.message] as JSON)
 		} finally {
-			if (deleteFile && temp.exists())
+			if (deleteFile && temp!=null && temp.exists())
 				temp.delete()
 		}
 	}
@@ -228,7 +233,11 @@ class StreamController {
 	def confirm() {
 		Stream stream = Stream.get(params.id)
 		File file = new File(params.file)
-		CSVImporter csv = new CSVImporter(file)
+
+		Map config = stream.streamConfig ? JSON.parse(stream.streamConfig) : [:]
+		List fields = config.fields ? config.fields : []
+
+		CSVImporter csv = new CSVImporter(file, fields)
 		Schema schema = csv.getSchema()
 		
 		[schema:schema, file:params.file, stream:stream]
@@ -237,6 +246,10 @@ class StreamController {
 	def confirmUpload() {
 		Stream stream = Stream.get(params.id)
 		File file = new File(params.file)
+
+		Map config = stream.streamConfig ? JSON.parse(stream.streamConfig) : [:]
+		List fields = config.fields ? config.fields : []
+
 		def format
 		def index
 		if(params.customFormat)
@@ -244,8 +257,7 @@ class StreamController {
 		else format = params.format
 		index = Integer.parseInt(params.timestampIndex)
 		try {
-			CSVImporter csv = new CSVImporter(file, index, format)
-			Schema schema = csv.getSchema()
+			CSVImporter csv = new CSVImporter(file, fields, index, format)
 			importCsv(csv, stream)
 		} catch (Exception e) {
 			flash.message = "The format of the timestamp is not correct"
@@ -253,25 +265,25 @@ class StreamController {
 		redirect(action:"show", id:params.id)
 	}
 	
-	private void importCsv(CSVImporter csv, Stream stream) {		
-		List<FeedFile> feedFiles = kafkaService.createFeedFilesFromCsv(csv, stream)
+	private void importCsv(CSVImporter csv, Stream stream) {
+		kafkaService.createFeedFilesFromCsv(csv, stream)
 		
 		// Autocreate the stream config based on fields in the csv schema
 		Map config = (stream.streamConfig ? JSON.parse(stream.streamConfig) : [:])
-		if (!config.fields || config.fields.isEmpty()) {
-			List fields = []
-			
-			// The primary timestamp column is implicit, so don't include it in streamConfig
-			for (int i=0;i<csv.schema.entries.length;i++) {
-				if (i!=csv.schema.timestampColumnIndex) {
-					CSVImporter.SchemaEntry e = csv.schema.entries[i]
+
+		List fields = []
+
+		// The primary timestamp column is implicit, so don't include it in streamConfig
+		for (int i=0;i<csv.schema.entries.length;i++) {
+			if (i!=csv.getSchema().timestampColumnIndex) {
+				CSVImporter.SchemaEntry e = csv.getSchema().entries[i]
+				if (e!=null)
 					fields << [name:e.name, type:e.type]
-				}
 			}
-			
-			config.fields = fields
-			stream.streamConfig = (config as JSON)
 		}
+
+		config.fields = fields
+		stream.streamConfig = (config as JSON)
 	}
 	
 	def deleteFeedFilesUpTo() {
